refactor: implement remote service and update movie repository
